# docs/openapi.json
{
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "scheme": "bearer",
        "type": "http"
      },
      "cookieAuth": {
        "in": "cookie",
        "name": "better-auth.session_token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "API de catálogo musical para o app Flutter. As rotas de autenticação (/api/auth/*) são gerenciadas pelo Better Auth com suporte simultâneo a Bearer Token e Cookie HttpOnly.",
    "title": "cardoso-sound-api",
    "version": "0.1.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/v1/playlists": {
      "get": {
        "description": "Recupera a listagem paginada de playlists pertencentes exclusivamente ao usuário ativo.",
        "operationId": "listPlaylists",
        "parameters": [
          {
            "description": "Número da página (>= 1)",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "maximum": 9007199254740991,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "description": "Quantidade de itens por página (1..100)",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "data": {
                      "description": "Lista de playlists da página solicitada",
                      "items": {
                        "additionalProperties": false,
                        "properties": {
                          "createdAt": {
                            "description": "Data de criação da playlist em formato ISO 8601 UTC",
                            "format": "date-time",
                            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$",
                            "type": "string"
                          },
                          "description": {
                            "description": "Descrição da playlist ou null",
                            "nullable": true,
                            "type": "string"
                          },
                          "id": {
                            "description": "Identificador único da playlist (UUID v4)",
                            "format": "uuid",
                            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                            "type": "string"
                          },
                          "name": {
                            "description": "Nome da playlist",
                            "type": "string"
                          },
                          "trackCount": {
                            "description": "Quantidade de faixas na playlist",
                            "maximum": 9007199254740991,
                            "minimum": 0,
                            "type": "integer"
                          },
                          "updatedAt": {
                            "description": "Data da última atualização da playlist em formato ISO 8601 UTC",
                            "format": "date-time",
                            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$",
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "name",
                          "description",
                          "trackCount",
                          "createdAt",
                          "updatedAt"
                        ],
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "meta": {
                      "additionalProperties": false,
                      "description": "Metadados de paginação da listagem",
                      "properties": {
                        "hasNext": {
                          "description": "Indica se existe uma próxima página",
                          "type": "boolean"
                        },
                        "hasPrev": {
                          "description": "Indica se existe uma página anterior",
                          "type": "boolean"
                        },
                        "limit": {
                          "description": "Limite de itens solicitados por página",
                          "maximum": 9007199254740991,
                          "minimum": 1,
                          "type": "integer"
                        },
                        "page": {
                          "description": "Página atual da listagem",
                          "maximum": 9007199254740991,
                          "minimum": 1,
                          "type": "integer"
                        },
                        "total": {
                          "description": "Total geral de itens encontrados",
                          "maximum": 9007199254740991,
                          "minimum": 0,
                          "type": "integer"
                        },
                        "totalPages": {
                          "description": "Total de páginas disponíveis",
                          "maximum": 9007199254740991,
                          "minimum": 1,
                          "type": "integer"
                        }
                      },
                      "required": [
                        "page",
                        "limit",
                        "total",
                        "totalPages",
                        "hasNext",
                        "hasPrev"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "data",
                    "meta"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "details": {
                      "description": "Detalhes adicionais ou issues de validação RFC 7807",
                      "nullable": true
                    },
                    "error": {
                      "description": "Identificador canônico do erro",
                      "type": "string"
                    },
                    "message": {
                      "description": "Mensagem descritiva da falha",
                      "type": "string"
                    },
                    "statusCode": {
                      "description": "Código de status HTTP",
                      "maximum": 9007199254740991,
                      "minimum": -9007199254740991,
                      "type": "integer"
                    }
                  },
                  "required": [
                    "statusCode",
                    "error",
                    "message",
                    "details"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "details": {
                      "description": "Detalhes adicionais ou issues de validação RFC 7807",
                      "nullable": true
                    },
                    "error": {
                      "description": "Identificador canônico do erro",
                      "type": "string"
                    },
                    "message": {
                      "description": "Mensagem descritiva da falha",
                      "type": "string"
                    },
                    "statusCode": {
                      "description": "Código de status HTTP",
                      "maximum": 9007199254740991,
                      "minimum": -9007199254740991,
                      "type": "integer"
                    }
                  },
                  "required": [
                    "statusCode",
                    "error",
                    "message",
                    "details"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "summary": "Lista as playlists do usuário autenticado",
        "tags": [
          "Library"
        ]
      },
      "post": {
        "description": "Cria uma nova playlist com nome e descrição opcional. Limite máximo de 50 playlists por usuário.",
        "operationId": "createPlaylist",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "description": {
                    "description": "Descrição opcional da playlist (máximo 500 caracteres)",
                    "maxLength": 500,
                    "nullable": true,
                    "type": "string"
                  },
                  "name": {
                    "description": "Nome da playlist (1..120 caracteres)",
                    "maxLength": 120,
                    "minLength": 1,
                    "type": "string"
                  }
                },
                "required": [
                  "name"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "createdAt": {
                      "description": "Data de criação da playlist em formato ISO 8601 UTC",
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$",
                      "type": "string"
                    },
                    "description": {
                      "description": "Descrição da playlist ou null",
                      "nullable": true,
                      "type": "string"
                    },
                    "id": {
                      "description": "Identificador único da playlist (UUID v4)",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                      "type": "string"
                    },
                    "name": {
                      "description": "Nome da playlist",
                      "type": "string"
                    },
                    "trackCount": {
                      "description": "Quantidade de faixas na playlist",
                      "maximum": 9007199254740991,
                      "minimum": 0,
                      "type": "integer"
                    },
                    "updatedAt": {
                      "description": "Data da última atualização da playlist em formato ISO 8601 UTC",
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$",
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "description",
                    "trackCount",
                    "createdAt",
                    "updatedAt"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "details": {
                      "description": "Detalhes adicionais ou issues de validação RFC 7807",
                      "nullable": true
                    },
                    "error": {
                      "description": "Identificador canônico do erro",
                      "type": "string"
                    },
                    "message": {
                      "description": "Mensagem descritiva da falha",
                      "type": "string"
                    },
                    "statusCode": {
                      "description": "Código de status HTTP",
                      "maximum": 9007199254740991,
                      "minimum": -9007199254740991,
                      "type": "integer"
                    }
                  },
                  "required": [
                    "statusCode",
                    "error",
                    "message",
                    "details"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "details": {
                      "description": "Detalhes adicionais ou issues de validação RFC 7807",
                      "nullable": true
                    },
                    "error": {
                      "description": "Identificador canônico do erro",
                      "type": "string"
                    },
                    "message": {
                      "description": "Mensagem descritiva da falha",
                      "type": "string"
                    },
                    "statusCode": {
                      "description": "Código de status HTTP",
                      "maximum": 9007199254740991,
                      "minimum": -9007199254740991,
                      "type": "integer"
                    }
                  },
                  "required": [
                    "statusCode",
                    "error",
                    "message",
                    "details"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "details": {
                      "description": "Detalhes adicionais ou issues de validação RFC 7807",
                      "nullable": true
                    },
                    "error": {
                      "description": "Identificador canônico do erro",
                      "type": "string"
                    },
                    "message": {
                      "description": "Mensagem descritiva da falha",
                      "type": "string"
                    },
                    "statusCode": {
                      "description": "Código de status HTTP",
                      "maximum": 9007199254740991,
                      "minimum": -9007199254740991,
                      "type": "integer"
                    }
                  },
                  "required": [
                    "statusCode",
                    "error",
                    "message",
                    "details"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "summary": "Cria uma nova playlist para o usuário autenticado",
        "tags": [
          "Library"
        ]
      }
    },
...
    "/api/v1/playlists/{id}/tracks/{trackId}": {
      "delete": {
        "description": "Remove o vínculo de uma faixa específica com a playlist. Responde 404 se a playlist ou a faixa não estiverem presentes.",
        "operationId": "removeTrackFromPlaylist",
        "parameters": [
          {
            "description": "Identificador único da playlist (UUID v4)",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          },
          {
            "description": "Identificador único da faixa musical (UUID v4)",
            "in": "path",
            "name": "trackId",
            "required": true,
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "content": {
              "application/json": {
                "schema": {
                  "description": "Faixa removida da playlist com sucesso sem conteúdo retornado"
                }
              }
            },
            "description": "Faixa removida da playlist com sucesso sem conteúdo retornado"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "details": {
                      "description": "Detalhes adicionais ou issues de validação RFC 7807",
                      "nullable": true
                    },
                    "error": {
                      "description": "Identificador canônico do erro",
                      "type": "string"
                    },
                    "message": {
                      "description": "Mensagem descritiva da falha",
                      "type": "string"
                    },
                    "statusCode": {
                      "description": "Código de status HTTP",
                      "maximum": 9007199254740991,
                      "minimum": -9007199254740991,
                      "type": "integer"
                    }
                  },
                  "required": [
                    "statusCode",
                    "error",
                    "message",
                    "details"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "details": {
                      "description": "Detalhes adicionais ou issues de validação RFC 7807",
                      "nullable": true
                    },
                    "error": {
                      "description": "Identificador canônico do erro",
                      "type": "string"
                    },
                    "message": {
                      "description": "Mensagem descritiva da falha",
                      "type": "string"
                    },
                    "statusCode": {
                      "description": "Código de status HTTP",
                      "maximum": 9007199254740991,
                      "minimum": -9007199254740991,
                      "type": "integer"
                    }
                  },
                  "required": [
                    "statusCode",
                    "error",
                    "message",
                    "details"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "details": {
                      "description": "Detalhes adicionais ou issues de validação RFC 7807",
                      "nullable": true
                    },
                    "error": {
                      "description": "Identificador canônico do erro",
                      "type": "string"
                    },
                    "message": {
                      "description": "Mensagem descritiva da falha",
                      "type": "string"
                    },
                    "statusCode": {
                      "description": "Código de status HTTP",
                      "maximum": 9007199254740991,
                      "minimum": -9007199254740991,
                      "type": "integer"
                    }
                  },
                  "required": [
                    "statusCode",
                    "error",
                    "message",
                    "details"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "summary": "Remove uma faixa musical de uma playlist",
        "tags": [
          "Library"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Local",
      "url": "http://localhost:3333"
    }
  ],
...
}